Move tendency related funcs into tendency.py
